EC2 preferred config
This change makes it a little clearer how to set up a cloud miner. 
When I ran through this the first time, I searched the marketplace for AMIs which didn't work out cleanly. The better approach is to just spin up the instance from the EC2 dashboard.
Additionally, I added some notes around using a t2.small instance. I've found this to be the 'happy place' for the miner after some testing.

If possible, the "Configure AWS Instance for Sync" section would benefit with these images as illustration. I can't add them outside of gitbook.

https://joeyhillercom-misc.s3-us-west-1.amazonaws.com/temp/Screen+Shot+2020-07-08+at+6.04.41+PM.png
https://joeyhillercom-misc.s3-us-west-1.amazonaws.com/temp/Screen+Shot+2020-07-08+at+6.05.02+PM.png